Renamed global header and added helpful links to README.
* Renamed Header for consistency.
* Added todos to README to document helpful links.
