+++ packages/auth/auth0-react/src/components/Header.tsx
@@ -1,0 +1,33 @@
+import { Link } from "react-router-dom";
+import { useAuth0 } from "@auth0/auth0-react";
+
+import LoginButton from "./LoginButton";
+import LogoutButton from "./LogoutButton";
+
+const Header = () => {
+  const { isAuthenticated } = useAuth0();
+  return (
+    <header>
+      <nav>
+        <ul>
+          <li>
+            <Link to="/">Home</Link>
+          </li>
+          <li>
+            <Link to="/about">About</Link>
+          </li>
+          <li>
+            <Link to="/admin">Admin</Link>
+          </li>
+        </ul>
+      </nav>
+
+      <div>
+        {!isAuthenticated && <LoginButton />}
+        {isAuthenticated && <LogoutButton />}
+      </div>
+    </header>
+  );
+};
+
+export default Header;
