--- packages/auth/auth0-react/src/components/header.global.tsx
@@ -1,21 +1,0 @@
-import { Link } from "react-router-dom";
-
-const Header = () => {
-  return (
-    <header>
-      <ul>
-        <li>
-          <Link to="/">Home</Link>
-        </li>
-        <li>
-          <Link to="/about">About</Link>
-        </li>
-        <li>
-          <Link to="/admin">Admin</Link>
-        </li>
-      </ul>
-    </header>
-  );
-};
-
-export default Header;
