+++ packages/auth/auth0-react/src/pages/errors/NotFound.tsx
@@ -1,7 +1,7 @@
 import PageProps from "../../types/PageProps";
 import { useDocumentTitle } from "../../hooks/useDocumentTitle";
 
-// See https://stackoverflow.com/a/75948984 for error typing
+// TODO: Add to README See https://stackoverflow.com/a/75948984 for error typing
 const NotFound = ({ title }: PageProps) => {
   const documentTitle = useDocumentTitle;
   documentTitle(title);
